Improve error handling architecture for the Quarkus server, utilize an error factory on the GUI client, and remove git patches.

+++ vcell-rest/src/main/java/org/vcell/restq/handlers/SolverResource.java
@@ -21,19 +21,19 @@ import org.apache.logging.log4j.Logger;
 import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
 import org.jboss.resteasy.reactive.RestForm;
+import org.vcell.restq.errors.exceptions.RuntimeWebException;
+import org.vcell.restq.errors.exceptions.UnprocessableContentWebException;
 import org.vcell.sbml.FiniteVolumeRunUtil;
 import org.vcell.sbml.vcell.SBMLExporter;
 import org.vcell.sbml.vcell.SBMLImporter;
 import org.w3c.www.http.HTTP;
 
-import javax.swing.text.html.Option;
 import java.beans.PropertyVetoException;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
-import java.util.Optional;
 
 @Path("/api/v1/solver")
 @RequestScoped
@@ -51,10 +51,12 @@ public class SolverResource {
     @Produces(MediaType.APPLICATION_OCTET_STREAM)
     public File retrieveFiniteVolumeInputFromSBMLSpatial(@RestForm File sbmlFile,
                                                          @RestForm @DefaultValue("5.0") double duration,
-                                                         @RestForm @DefaultValue("0.1") double output_time_step) throws IOException {
-        File workingDir = Files.createTempDirectory("vcell-").toFile();
-        File zipFile = Files.createTempFile("finite-volume", ".zip").toFile();
+                                                         @RestForm @DefaultValue("0.1") double output_time_step) throws UnprocessableContentWebException {
+        File workingDir = null;
+        File zipFile = null;
         try {
+            workingDir = Files.createTempDirectory("vcell-").toFile();
+            zipFile = Files.createTempFile("finite-volume", ".zip").toFile();
             sbmlToFiniteVolumeInput(sbmlFile, workingDir, duration, output_time_step);
             ZipFile zip = new ZipFile(zipFile);
             for (File file : workingDir.listFiles()) {
@@ -62,14 +64,17 @@ public class SolverResource {
             }
             zip.close();
             return zipFile;
-        }catch (Exception e){
-            lg.error(e);
-            throw new WebApplicationException("Error processing spatial model", HTTP.INTERNAL_SERVER_ERROR);
+        } catch (PropertyVetoException | MappingException | SolverException | ExpressionException e){
+            throw new UnprocessableContentWebException(e.getMessage(), e);
+        } catch (VCLoggerException | IOException e){
+            throw new RuntimeWebException("Error processing spatial model: " + e.getMessage(), e);
         } finally {
-            for (File file: workingDir.listFiles()){
-                file.delete();
+            if (workingDir != null && zipFile != null){
+                for (File file: workingDir.listFiles()){
+                    file.delete();
+                }
+                workingDir.delete();
             }
-            workingDir.delete();
         }
     }
 
@@ -104,10 +109,12 @@ public class SolverResource {
     @Parameter(name = "simulation_name", description = "name of simulation in VCML file")
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces(MediaType.APPLICATION_OCTET_STREAM)
-    public File retrieveFiniteVolumeInputFromVCML(@RestForm File vcmlFile, @RestForm String simulation_name) throws IOException {
-        File workingDir = Files.createTempDirectory("vcell-").toFile();
-        File zipFile = Files.createTempFile("finite-volume", ".zip").toFile();
+    public File retrieveFiniteVolumeInputFromVCML(@RestForm File vcmlFile, @RestForm String simulation_name) throws UnprocessableContentWebException {
+        File workingDir = null;
+        File zipFile = null;
         try {
+            workingDir = Files.createTempDirectory("vcell-").toFile();
+            zipFile = Files.createTempFile("finite-volume", ".zip").toFile();
             vcmlToFiniteVolumeInput(vcmlFile, simulation_name, workingDir);
             ZipFile zip = new ZipFile(zipFile);
             for (File file : workingDir.listFiles()) {
@@ -115,14 +122,17 @@ public class SolverResource {
             }
             zip.close();
             return zipFile;
-        }catch (XmlParseException | MappingException | SolverException | ExpressionException e){
-            lg.error(e);
-            throw new WebApplicationException("Error processing spatial model: "+e.getMessage(), HTTP.NOT_ACCEPTABLE);
+        } catch (XmlParseException | MappingException | SolverException | ExpressionException e){
+            throw new UnprocessableContentWebException("Error processing spatial model: "+e.getMessage(), e);
+        } catch (IOException e){
+            throw new RuntimeWebException("IO Error on server: " + e.getMessage(), e);
         } finally {
-            for (File file: workingDir.listFiles()){
-                file.delete();
+            if (workingDir != null && zipFile != null){
+                for (File file: workingDir.listFiles()){
+                    file.delete();
+                }
+                workingDir.delete();
             }
-            workingDir.delete();
         }
     }
